// src/app/Shared/Models/Login.ts
// Generated by https://quicktype.io

export interface Login {
    userName: string;
    password: string;
}

// Generated by https://quicktype.io

export interface LoginResponse {
    username:  string;
    jwtToken:  string;
    expiresIn: number;
}

// Generated by https://quicktype.io

export interface User {
    name: string;
    role: string;
    nbf:  number;
    exp:  number;
    iat:  number;
}